Refactor Navbar and Footer components to global directory and update styles
