+++ src/App.tsx
@@ -1,5 +1,5 @@
-import Footer from './components/Footer';
-import Navbar from './components/Navbar';
+import Footer from './components/global/Footer';
+import Navbar from './components/global/Navbar';
 import Homepage from './pages/Homepage';
 import Signup from './pages/Signup';
 import { Routes, Route } from 'react-router';
